Player data
- Unit creation substract resources
- On unit production cancel player got resources back

+++ Assets/Scripts/Game Logic/Units/UnitInfo.cs
@@ -19,9 +19,9 @@ public class UnitInfo : ScriptableObject
     [SerializeField] private float moveSpeed = 0;
 
     [Header("Unit cost")]
-    [SerializeField] private float plasm = 0;
-    [SerializeField] private float titan = 0;
-    [SerializeField] private float chips = 0;
+    [SerializeField] private int plasm = 0;
+    [SerializeField] private int titan = 0;
+    [SerializeField] private int chips = 0;
 
     public int Hp => hp;
     public int Damage => damage;
@@ -36,7 +36,7 @@ public class UnitInfo : ScriptableObject
     public float RotationSpeed => rotationSpeed;
     public float MoveSpeed => moveSpeed;
 
-    public float Plasm => plasm;
-    public float Titan => titan;
-    public float Chips => chips;
+    public int Plasm => plasm;
+    public int Titan => titan;
+    public int Chips => chips;
 }
